requirements.txt
boto3==1.16.51
Django==3.1.5
django-cors-headers==3.6.0
django-filter==2.4.0
django-imagekit==4.0.2
django-heroku==0.3.1
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
django-ses==1.0.3
django-storages==1.11.1
drf-yasg==1.20.0
gunicorn==20.0.4
onesignal-sdk==2.0.0
phonenumbers==8.12.15
Pillow==8.1.0
psycopg2==2.8.6
PyJWT==1.7.1
twilio==6.51.0
redis==3.5.3
requests~=2.25.1